Refactor : useManageFilter 제작
