+++ src/components/feature/Filter/Filter.tsx
@@ -1,4 +1,4 @@
-import { MouseEventHandler, RefObject, ReactNode, createContext, useState, useContext } from 'react';
+import { MouseEventHandler, RefObject, useContext } from 'react';
 
 import { Address as AddressType } from '@/apis/common.type';
 import Badge from '@/components/common/Badge/Badge';
@@ -7,54 +7,29 @@ import styles from '@/components/feature/Filter/Filter.module.scss';
 import ScrollMenu from '@/components/feature/Filter/ScrollMenu/ScrollMenu';
 import { ReactComponent as CloseButton } from '@/public/svgs/closeButton.svg';
 
-type FilterContextProps = {
-  isOpen: boolean;
-  close: () => void;
-  open: () => void;
-};
-
-export const FilterContext = createContext<FilterContextProps | null>(null);
-export const useFilter = () => {
-  const context = useContext(FilterContext);
-  if (!context) throw new Error('root 로 감싸기');
-  return context;
-};
-export function FilterProvider({ children }: { children: ReactNode }) {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const open = () => {
-    setIsOpen(true);
-  };
-  const close = () => {
-    setIsOpen(false);
-  };
-  // eslint-disable-next-line react/jsx-no-constructed-context-values
-  return <FilterContext.Provider value={{ isOpen, open, close }}>{children}</FilterContext.Provider>;
-}
+import { FilterContext } from './FilterContext';
 
 type FilterProps = {
-  // isFilterShow?: boolean;
-  filterRef: RefObject<HTMLDivElement>;
-  // setShowFilter: Dispatch<SetStateAction<boolean>>;
   scrollMenuList: AddressType[];
   handleMenuClick: MouseEventHandler;
-  clickedAddress: AddressType[];
+  clickedAddress: AddressType[] | undefined;
   handleResetBtnClick: MouseEventHandler;
   handleApplyBtnClick?: MouseEventHandler;
+  filterRef: RefObject<HTMLDivElement>;
+  // handleInputChange : SubmitEvent;
 };
+
 /**
  * Filter 컴포넌트
- * param isFilterShow 외부에서 필터 open 버튼 클릭시 true로 들어가도록 넣어주세요 (state 관리), requried ; boolean
- * param filterRef 필터에 걸 ref. useClickFilterOutsidegnrdml filterRef 가져다 쓰면 됨, required ; RefObject<HTMLDivElement>
- * param setShowFilter filter를 켤지 끌 지 정하는 setter, required ; Dispatch<React.SetStateAction<boolean>>;
  * @param scrollMenuList 스크롤 메뉴에 들어갈 list, 우리 프로젝트에서는 constatns/addressList required; sring[]
  * @param handleMenuClick 메뉴를 클릭했을 때의 핸들러 . required ; MouseEventHanler
  * @param clickedAddress 클릭된 address들을 모아둔 List. required ; string[]
  * @param handleResetBtnClick 초기화 버튼 클릭시 핸들러. clickedAddress를 초기화해주고 input값도 초기화해줌. required; MouseEventHandler
  * @param handleApplyBtnClick 적용하기 버튼 클릭 시 동작할 handler ; MouseEventHandler
+ * @param filterRe 필터에 걸어둔 ref
  */
 
-export function Filter({
+export default function Filter({
   scrollMenuList,
   handleMenuClick,
   clickedAddress,
@@ -63,9 +38,11 @@ export function Filter({
   filterRef
 }: FilterProps) {
   const filterContext = useContext(FilterContext);
+
   if (!filterContext) {
     return null; // 로딩처리
   }
+
   const { close } = filterContext;
   return (
     <div className={styles.container} ref={filterRef}>
@@ -78,7 +55,7 @@ export function Filter({
         <p>위치</p>
         <ScrollMenu list={scrollMenuList} handleClick={handleMenuClick} />
         <section className={styles.badgeList}>
-          {clickedAddress.map((item) => (
+          {clickedAddress?.map((item) => (
             <div key={item}>
               <Badge title={item} color="red" closeBtn />
             </div>
