+++ src/components/feature/Filter/FilterContext.tsx
@@ -1,0 +1,26 @@
+import { createContext, useContext, useState, ReactNode } from 'react';
+
+type FilterContextProps = {
+  isOpen: boolean;
+  close: () => void;
+  open: () => void;
+};
+
+export const FilterContext = createContext<FilterContextProps | null>(null);
+export const useFilter = () => {
+  const context = useContext(FilterContext);
+  if (!context) throw new Error('root 로 감싸기');
+  return context;
+};
+export function FilterProvider({ children }: { children: ReactNode }) {
+  const [isOpen, setIsOpen] = useState(false);
+
+  const open = () => {
+    setIsOpen(true);
+  };
+  const close = () => {
+    setIsOpen(false);
+  };
+  // eslint-disable-next-line react/jsx-no-constructed-context-values
+  return <FilterContext.Provider value={{ isOpen, open, close }}>{children}</FilterContext.Provider>;
+}
